Initial commit.
This is "ODK-Core", an experimental project to provide the executable
workflows and seeding script from the ODK as an independent Python
package.

This commit is a very rough draft, containing the original ODK templates
and a slightly modified seeding script (modified so that it is splitted
in a few modules, rather than being in a single big monolithic module).