Centralize the replacement of stub IDs by product objects.
"Stub IDs" refers to the possibility of listing the products of a
product group (e.g., imports) as a simple list of IDs, rather than as an
explicit list of products. That is, for example, with:

```
import_group:
  ids:
    - my-import
    - my-other-import
```

This is automatically transformed into:

```
import_group:
  products:
    - id: my-import
    - id: my-other-import
```

But the fact that we are doing the transformation _after_ the dictionary
has been parsed into a OntologyProject is annoying, because it means
that we must define the `products` field of the group as being
_optional_. This is turn means that, even after the above transformation
has been done, MyPy will still forever consider that the `products`
field may contain a None value, and will force us to add needless guard
conditions.

So here, we change the way we realise that transformation: we transform
the dictionary before parsing it into an OntologyProject object, rather
than transforming the OntologyProject. That means that we no longer need
to artificially declare the `products` field as optional.

This is also cleaner as it means the model classes are no longer
polluted by those `ids` fields which serve no other purpose than to
support the "Stub IDs" syntax.