Uncouple the workflows from the ODK image.
The standard workflows (as implemented by the template-derived Makefile)
should _not_ assume that they are run from within the ODK Docker image.

Currently, this assumption is made for two types of things:

(1) The workflows sometimes need to access some resources found in the
ODK image; when they do that, the path to these resources is hardcoded.

(2) The workflows sometimes call a script that is specific to the ODK
image, and will not exist anywhere else (e.g. the `check-rdfxml`
script).

To fix the first case, we introduce an environment variable called
ODK_RESOURCES_DIR, which may be set to point to a different directory
depending on how the workflows are run. For example, when a workflow is
run from within the ODK Docker image, that variable may be set to point
to the `/tools/resources`, which is the dedicated resources directory in
the Docker image. When _not_ using the Docker image, it is up to the
ODK Runner (whatever that is) to set up that variable to point to the
appropriate place.

To fix the second case, we introduce a new tool in the ODK Core package,
`odk-helper`, that will be intended to provide all the features that are
currently provided by ODK-specific helper scripts.