Added (experimental) `install` command.
Add a new user command `install` to create a basic runtime environment
to run ODK workflows.

The command creates a directory containing all the ODK-specific tools
(ROBOT, DOSDP-Tools, SSSOM-CLI, etc.), the ODK ROBOT plugins (the `odk`
plugin and the `sssom` plugin) and the OBO extended prefix map, then
provides a "activation" file that can be sourced to make the tools and
resources available.

The idea is that the user can do:

  $ odk install myenv
  $ source myenv/bin/activate-odk-environment.sh

and from that point on the current shell is ready to run ODK workflows.

Limitations:

* This will never work on Windows.
* Non-ODK-specific tools (e.g. Make, Git, plus all the "classic" Unix
  tools like sed, grep, etc.) are not included and are assumed to be
  already available on the system.