Remove the distinction between "tools" and "files".
When installing an ODK environment, do not treat "tools" distinctly from
the more general "files". For all files (whether they represent a "tool"
or not), check whether the file is already available, and if it is then
we only re-install it under the `--force` option.

The only special thing about "tools" is that a tool is considered to be
available if it is in the PATH, even if it is not in the actual bin
directory of the target environment.