Pass Java options to Java programs.
The ODK Runner collects all options for the Java virtual machine into
two variables: ROBOT_JAVA_ARGS and JAVA_OPTS (they are set to the same
value).

Make sure the launchers for Java tools pass the $JAVA_OPTS variable to
the virtual machine, so that the options are not ignored.