Improve the activation script.
The current "activate-odk-environment.sh" script is rudimentary and does
not allow to "deactivate" the environment (the only to do that is to
leave the shell from which the script was sourced).

We replace it by a more complete script heavily inspired from the
"activate" script generated by Python virtual environment tools such as
virtualenv or uv.