Add a helper command for downloading files.
This commit adds a new command to the `odk-helper` tool: `download`. The
new command is intended to be used from within ODK workflows to download
files (e.g. when mirroring an ontology, or when refreshing a remotely
sourced component).

The aim is to simplify the standard Makefile by removing from it all the
suff needed to ensure that the rules that depend on a downloaded file
are not needlessly triggered when we download a file that has, in fact,
not changed since the last time it was downloaded.

The `odk-helper download` command takes care of that by keeping track,
for every downloaded file, of up to three pieces of information:

* the time at which the file was downloaded;
* the ETag returned by the server for that file (if any);
* the SHA-256 checksum of the downloaded file.

When we ask to download the same file again, the command uses the time
of last download and the ETag (if available) to specifically ask the
server for a newer version of the file, thereby avoiding download
completely if the server is able to honor the If-Modified-Since and/or
If-None-Match headers. If we do download the file, then we first write
it into a temporary location, compute its SHA-256 checksum, and compare
it with the recorded checksum from the last download; if the checksums
match, then the newly downloaded file is in fact not new and the
temporary file is deleted without ever touching the existing file.

In addition, the command can automatically:

* uncompress a .gz or .bz2 file;
* try to download a compressed version of a remote resource first (by
  appending `.gz` to the requested URL), then fallback to the original
  URL.