Refactor all downloading rules.
Now that we have a `odk-helper download` command, we can use it to
overhaul the way we download mirrors and remotely sourced components in
the standard Makefile.

Basically, to download and prepare a mirror M, we have two rules:

(1) A "phony" rule `download-mirror-M`. That rule is responsible for
actually downloading the remote file. It does so with the `odk-helper
download` command. The newly downloaded file is expected to be written
into `$(TMPDIR)/download-mirror-M.owl`, _unless_ the file has not been
changed since the last time it was downloaded (in which case nothing
will be written at that location). It is also in that rule that we deal
with the `use_base` and `use_gzipped` options.

(2) A "normal" rule `$(MIRRORDIR)/M.owl`, which depends on the first
rule. That rule should convert the downloaded file into the final mirror
file, by running `robot remove` (if `make_base` is enabled for the
mirror) and `robot convert`. Importantly, that rule must check
beforehand whether the `$(TMPDIR)/download-mirror-M.owl` file exists,
and do nothing if it does not (which will be the case when the remote
file has not changed since the last time it was downloaded).

The same principles applies for downloading remotely sourced components.